NeuralFin-Django/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
cssselect==1.2.0
DateTime==5.1
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
fake-useragent==1.1.3
feedparser==6.0.10
filelock==3.9.0
finvizfinance==0.14.5
frozendict==2.3.6
gunicorn==20.1.0
html5lib==1.1
huggingface-hub==0.12.1
ibm-cloud-sdk-core==3.16.5
idna==3.4
importlib-metadata==6.1.0
investpy==1.0.8
lxml==4.9.2
multitasking==0.0.11
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parse==1.19.0
psycopg2-binary==2.9.5
pycparser==2.21
pyee==8.2.2
Pygments==2.14.0
PyJWT==2.6.0
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-futures==1.0.0
requests-html==0.10.0
sentencepiece==0.1.97
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
Unidecode==1.3.6
urllib3==1.26.14
w3lib==2.1.1
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yahooquery==2.3.1
yfinance==0.2.13
zipp==3.15.0
zope.interface==6.0
